WebSockets/tests/submissions/Opera/constructor/004.html

<!doctype html>
<title>WebSockets: new WebSocket(url, invalid protocol)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../constants.js></script>
<div id=log></div>
<script>
function checkException(e) {
  assertInstanceOf(e, DOMException);
  assert_equals(e.code, e.SYNTAX_ERR);
}
// empty string
assertThrows(function(){new WebSocket("ws://"+location.hostname+"/", "")}, checkException);

// chars below U+0020 except U+0000; U+0000 is tested in a separate test
for (var i = 1; i < 0x20; ++i) {
  assertThrows(function(){new WebSocket("ws://"+location.hostname+"/", "a"+String.fromCharCode(i)+"b")}, checkException, 'char code '+i);
}
// some chars above U+007E
for (var i = 0x7F; i < 0x100; ++i) {
  assertThrows(function(){new WebSocket("ws://"+location.hostname+"/", "a"+String.fromCharCode(i)+"b")}, checkException, 'char code '+i);
}
end();
</script>
